+++ src/main/java/com/wanted/jaringoby/session/applications/LoginService.java
@@ -3,7 +3,7 @@ package com.wanted.jaringoby.session.applications;
 import com.wanted.jaringoby.common.utils.JwtUtil;
 import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.customer.exceptions.CustomerNotFoundException;
-import com.wanted.jaringoby.customer.exceptions.CustomerPasswordMismatchedException;
+import com.wanted.jaringoby.session.exceptions.CustomerPasswordMismatchedException;
 import com.wanted.jaringoby.customer.models.customer.Customer;
 import com.wanted.jaringoby.customer.repositories.CustomerRepository;
 import com.wanted.jaringoby.session.dtos.LoginRequestDto;
